First Working Prototype
That was quicker than I thought it would be :P

So this is really early, but it works. You can drag the nodes around,
and drag between 2 circles to attempt to link them. If it works (meaning they
are the same color, and going Input <> Output), a line
will be displayed which will stay connected when you drag the nodes, and
auto-disconnects when you connect to an input that already has a link.

Clicking "Render" will render the image (assuming your links are valid),
and dump it as "firstone.gif". Proper export settings will come
eventually :P

+Added basic Node functionality as described above
+Added Render button
~Changed Node size.